changed rating to weed leafs and made them clickable.

+++ app/src/main/java/com/mdksolutions/flowr/viewmodel/ProductDetailViewModel.kt
@@ -156,7 +156,7 @@ class ProductDetailViewModel : ViewModel() {
     // ─────────────────────────────────────────────────────────────
 
     /** Begin editing a specific review */
-    fun startEditingReview(reviewId: String, currentText: String, currentRating: Int) {
+    fun startEditingReview(reviewId: String, currentText: String, currentRating: Double) {
         _uiState.update {
             it.copy(
                 editingReviewId = reviewId,
